# Conflicts:
#	src/service/MakePolicyService.java

+++ src/enumeration/calculationFormula/homeFormula/HomeCompensation.java
@@ -1,7 +1,7 @@
 package enumeration.calculationFormula.homeFormula;
 
 public enum HomeCompensation {
-	
+
 	MoreThan450M(450_000_000, 500_000_000, "4.5억원 이상 5억원 이하"),
 	MoreThan400M(400_000_000, MoreThan450M.getMinAmount()-1, "4억원 이상"),
 	MoreThan350M(350_000_000, MoreThan400M.getMinAmount()-1, "3.5억원 이상"),
@@ -13,21 +13,21 @@ public enum HomeCompensation {
 	MoreThan50M(50_000_000, MoreThan100M.getMinAmount()-1, "5천만원 이상"),
 	MoreThan10M(10_000_000, MoreThan50M.getMinAmount()-1, "천만원 이상");
 
-	private final long minAmount;
-	private final long maxAmount;
+	private final int minAmount;
+	private final int maxAmount;
 	private final String name;
 	
-	HomeCompensation(long minAmount, long maxAmount, String name){
+	HomeCompensation(int minAmount, int maxAmount, String name){
 		this.minAmount = minAmount;
 		this.maxAmount = maxAmount;
 		this.name = name;	
 	}
 
-	public long getMinAmount() {
+	public int getMinAmount() {
 		return minAmount;
 	}
 
-	public long getMaxAmount() {
+	public int getMaxAmount() {
 		return maxAmount;
 	}
